refactor: move product routes to "./routes/products"
